+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 28/3-2024
+
+- Add search fields to some models in django admin.
+
 ## 20/6-2023
 
 - It is now possible to add moves to a pokémon from the UI.